+++ R/HDI.R
@@ -37,6 +37,7 @@
 #' @references Kruschke JK. Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan. 2nd edition. Academic Press, 2015
 #'
 #' @examples
+#' \dontrun{
 #' if (require("rstanarm")) {
 #'   fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
 #'   hdi(fit)
@@ -59,7 +60,7 @@
 #'   # posterior distribution values, which - after being exponentiated -
 #'   # are between .8 and 1.25 (about -.22 and .22 on linear scale)
 #'   rope(fit, rope = c(.8, 1.25), trans = exp)
-#' }
+#' }}
 #'
 #' @importFrom tibble as_tibble rownames_to_column
 #' @importFrom purrr map_dbl map_df